Get rid of PassManager for MoveConstructors
